% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_balance.r
\name{balance_test_multi_var}
\alias{balance_test_multi_var}
\title{Blance Test by F-test of Overall Significance in Linear Regression}
\usage{
balance_test_multi_var(
  covariate = NULL,
  treat = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = NULL,
  ctrl = NULL,
  subset = NULL,
  weights = NULL,
  cluster = NULL,
  ...
)
}
\arguments{
\item{covariate}{character vector of covariates}

\item{treat}{character of variable name of experimental arm}

\item{data}{data.frame/tibble you want to use}

\item{treat_levels}{order of experimental arms}

\item{treat_labels}{labels of experimental arms}

\item{ctrl}{specify control arm if you want change}

\item{subset}{subset condition of data}

\item{weights}{weight variable}

\item{cluster}{cluster variable}

\item{\dots}{pass to \code{estimatr::lm_robust()}}
}
\value{
A data frame with \code{balance_test} class
}
\description{
This function performs a balance test of the treatment
with multiple covariates.
The balance test calculates the mean of each group
and the F-test by linear regression analysis.
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- create_RCTtoolbox(
  atest + avacc ~ treat,
  ~ age + educ,
  RubellaNudge,
  LETTERS[1:7]
)

rct$balance(subset = coupon == 1)$result

}


}
