% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optRCTtool.r
\name{clear_optRCTtool}
\alias{clear_optRCTtool}
\title{Clear Global Options for RCTtoolbox}
\usage{
clear_optRCTtool()
}
\description{
Clear Global Options for RCTtoolbox
}
\examples{
\dontrun{
clear_optRCTtool()
options()[grep("RCTtool", names(options()))]
}

}
