% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox-create.r
\name{create_RCTtoolbox}
\alias{create_RCTtoolbox}
\title{Create New Toolbox}
\usage{
create_RCTtoolbox(
  baseline = NULL,
  covariate = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = treat_levels
)
}
\arguments{
\item{baseline}{two-sided formula with
one or more outcome variables on LHS and
one treatment variable on RHS.}

\item{covariate}{(list of) one-sided formulas with
covariates used in regression and balance test on RHS.}

\item{data}{data.frame/tibble object you want to use.}

\item{treat_levels}{character vector. Levels of experimental arms.
The first element is control arm.}

\item{treat_labels}{character vector.
Labels of experimental arms corresponding to \code{treat_levels}.
Default is to use \code{treat_levels} as labels.}
}
\value{
R6 object with "RCTtoolbox" class.
The returned object has following fields and methods
\describe{
\item{\code{data}}{Field.
Store data.frame/tibble passed to \code{data} argument.
You can access via \code{$data}}
\item{\code{print()}}{Method.
Print information about the returned object.
Run \code{$print()}}
\item{\code{ttest()}}{Method.
Implement two-sided t-test or permutation test.
Run \code{$ttest()}.
See \link{ttest} for help.}
\item{\code{power()}}{Method.
Implement power analysis.
Run \code{$power()}.
See \link{power} for help.}
\item{\code{balance()}}{Method.
Implement balance test.
Run \code{$balance()}.
See \link{balance} for help.}
\item{\code{lm()}}{Method.
Estimate linear model.
Run \code{$lm()}.
See \link{lm} for help.}
}
}
\description{
Create an R6 class object called "RCTtoolbox".
}
\section{Private fields}{

The created R6 object has also private fields.
You cannot access from the outside.
However, methods of RCTtoolbox use private fields.
A list of private fields is as follows:
\describe{
\item{formula.yd}{List of two-sided formula
\code{outcome ~ treatment}.
If you pass \code{y1 + y2 ~ d} to
the argument \code{baseline},
\code{formula.yd} refers to
\code{list(y1 ~ d, y2 ~ d)}.}
\item{formula.x}{Lists of one-sided formula
with covariates on RHS
which is passed to the argument \code{covariate}.}
\item{yvec}{Character vector of outcome variables.}
\item{xvec}{Character vector of covariates.}
\item{dvec}{Character of treatment.}
\item{dvec.levels}{Character vector
passed to the argument \code{treat_levels}.}
\item{dvec.labels}{Character vector
passed to the argument \code{treat_labels}.}
}
}

\examples{
\dontrun{
data(RubellaNudge)
create_RCTtoolbox(
  baseline = itest + ivacc ~ treat,
  covariate = ~ age + educ,
  data = RubellaNudge,
  treat_levels = LETTERS[1:7]
)
}


}
