% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.r
\name{expand_basic_mod}
\alias{expand_basic_mod}
\title{Generate expanded version of basic model}
\usage{
expand_basic_mod(y, d)
}
\arguments{
\item{y}{a string vector with outcome variables}

\item{d}{a string of treatment variable}
}
\description{
Generate expanded version of basic model
}
