% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expinfo.r
\name{expinfo}
\alias{expinfo}
\title{Create List of Experimental Information}
\usage{
expinfo(d, ctrl = NULL)
}
\arguments{
\item{d}{treatment vector.}

\item{ctrl}{character of the control.
If NULL, the first level of variable \code{d} is control.}
}
\description{
Create List of Experimental Information
}
