% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_test.r
\name{ftest}
\alias{ftest}
\title{F-test of Overall Siginificance in Linear Regression}
\usage{
ftest(mod, data)
}
\arguments{
\item{mod}{two-sided formula.
The covariate is specified in left-hand side of formula.
The treatment is specified in right-hand side of formula.}

\item{data}{data which you want to use.}
}
\description{
F-test of Overall Siginificance in Linear Regression
}
