% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.r
\name{genmod}
\alias{genmod}
\title{Generate List of Regression Models}
\usage{
genmod(basemod, xmod = NULL, include_onlyd = TRUE)
}
\arguments{
\item{basemod}{list of baseline model.
Baseline model is specified by \code{y ~ d}
where \code{y} is outcome, and \code{d} is treatments.}

\item{xmod}{list of covariate model.
Covariate model is specified by one-sided formula like \code{~ x1 + x2}.}

\item{include_onlyd}{logical.
Whether to estimate a model without covariates.}
}
\description{
Generate List of Regression Models
}
