% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-diff-test.r
\name{mean_diff_test}
\alias{mean_diff_test}
\title{T-Test for RCTs with Multiple Experimental Arms}
\usage{
mean_diff_test(outcome, arms, ctrl, data)
}
\arguments{
\item{outcome}{a string vector with outcome variables.
If missing, try to find out from \code{option("RCTtool.outcome")}.}

\item{arms}{a string of treatment variable.
If missing, try to find out from \code{option("RCTtool.arms")}.}

\item{ctrl}{a string of label of the control.
If missing, try to find out from \code{option("RCTtool.control")}.}

\item{data}{data object you want to use.}
}
\value{
tibble frame containing with five variables.
\code{treat} is a factor of treatments (baseline is the control).
\code{mean} and \code{se} is a mean and its standard error of each group.
\code{effect} is a difference of mean between each group and the control.
\code{p.value} is a p-value of t-test.
}
\description{
This function performs a t-test to test the null hypothesis
that there is no mean difference between the two groups.
When there are three or more treatment groups,
this function tests the difference of mean between
the control group specified by the user with the \code{ctrl} argument
and each group (including the control group).
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d1 <- sample(c("A", "B", "C"), size = n, replace = TRUE)
d2 <- ifelse(d1 == "A", "control", "treat")
ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- 0.2 * x1 + 0.5 * x2
y1 <- ifelse(d1 == "A", ya, ifelse(d1 == "B", yb, yc))
y2 <- ifelse(y1 > 1, 1, 0)
dt <- data.frame(
  outcome = y1,
  bin_outcome = y2,
  treat = d1,
  bin_treat = d2,
  x1,
  x2
)

# t-test
set_optRCTtool(outcome ~ treat, ~ x1 + x2, dt, "A")
mean_diff_test(data = dt)
mean_diff_test(data = subset(dt, treat != "B"))
mean_diff_test("bin_outcome", data = dt)
mean_diff_test(arms = "bin_treat", ctrl = "control", data = dt)

set_optRCTtool(bin_outcome ~ treat, data = dt, ctrl = "A")
mean_diff_test(data = dt)

}
