% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_RCTtool.r
\name{new_RCTtool}
\alias{new_RCTtool}
\title{Create New Toolbox}
\usage{
new_RCTtool(baseline = NULL, covariate = NULL, data = NULL, ctrl = NULL)
}
\arguments{
\item{baseline}{two-sided formula with
outcome variables on LHS and treatment variable on RHS.}

\item{covariate}{(list of) one-sided formulas with
covariates used in regression and balance test on RHS.}

\item{data}{data.frame/tibble object you want to use.}

\item{ctrl}{string of the control arm.}
}
\description{
Create an R6 class object called "RCT toolbox"
for consistent RCT data analysis.
The created object has method functions
such as t-test and regression analysis.
}
\examples{
\dontrun{
data(RubellaNudge)
new_RCTtool(
  baseline = itest + ivacc ~ treat,
  covariate = ~ age + educ,
  data = RubellaNudge,
  ctrl = "A"
)
}


}
