% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optRCTtool.r
\name{optRCTtool}
\alias{optRCTtool}
\title{Manipulation Global Options for RCTtoolbox}
\usage{
optRCTtool(..., clear = FALSE)
}
\arguments{
\item{\dots}{Specify \code{option_name = option}}

\item{clear}{logical (default is FALSE).
Whether to clear global options.}
}
\description{
Manipulation Global Options for RCTtoolbox
}
\examples{
\dontrun{
optRCTtool(
  RCTtool.outcome = c("y1", "y2"),
  RCTtool.arms = ~ treatment
)
getOption("RCTtool.outcome")
getOption("RCTtool.arms")
getOption("RCTtool.nothing") # not register by .onLoad
optRCTtool(clear = TRUE)
}

}
