% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.r
\name{power_analysis}
\alias{power_analysis}
\title{Power Analysis of Two Sided T-Test for RCT with Multiple Arms}
\usage{
power_analysis(
  mod,
  data,
  base = NULL,
  d = NULL,
  alpha = NULL,
  power = NULL,
  unstd_effect = TRUE,
  std_dev = 0.5
)
}
\arguments{
\item{mod}{formula. You can pass an one-sided formula \code{~ d} or
two-sided formula \code{y ~ d} where
\code{y} is outcome and \code{d} is treatments.}

\item{data}{data which you want to use.}

\item{base}{character. Name of the control.
If NULL (default), the first level of the variable \code{d} is control.}

\item{d}{numeric. Effect size.}

\item{alpha}{numeric. Significant level.}

\item{power}{numeric. Power.}

\item{unstd_effect}{logical.
Whether to calculate non-standarized effect.
Default is TRUE.}

\item{std_dev}{numeric.
The standard deviation used when calculating the destandardized effect.
If an outcome variable is specified,
this argument is ignored and the actual standard deviation is used.
Default is 0.5.}
}
\value{
tibble (and data.frame) with class "power.analysis"
}
\description{
This function performs a two-sided t-test power analysis
given the sample sizes of the treatment and control groups.
Calculate the value that is NULL among the effect size, significance level,
and power.
For the effect size, the non-standardized effect
(absolute value of the difference between the average outcomes)
is also calculated.
If the outcome variable is specified in the \code{mod} argument,
the square root of the mean of the outcome variances of the two group
is the standard deviation.
If no outcome variable is specified,
the standard deviation specified by the std_dev argument is used.
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))
dt <- data.frame(y, d, x1, x2)

# power analysis with variance assumption
power_analysis(~d, dt, alpha = 0.05, power = 0.8, std_dev = 0.2)

# power analysis with known variance
power_analysis(y ~ d, dt, alpha = 0.05, power = 0.8)


}
