% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_lm.r
\name{rct_lm}
\alias{rct_lm}
\title{Fitting Models by OLS}
\usage{
rct_lm(outcome, arms, ctrl, xmod, data, onlydmod = TRUE, subset_outcome, ...)
}
\arguments{
\item{outcome}{a string vector with outcome variables.
If missing, try to find out from \code{options("RCTtool.outcome")}.}

\item{arms}{a string of treatment variable.
If missing, try to find out from \code{options("RCTtool.arms")}.}

\item{ctrl}{a string of label of the control.
If missing, try to find out from \code{options("RCTtool.control")}.}

\item{xmod}{list of covariate model.
Covariate model is specified by one-sided formula like \code{~ x1 + x2}.
If missing, try to find out from \code{options("RCTtool.xmod")}.}

\item{data}{data which you want to use.}

\item{onlydmod}{logical (default is TRUE).
Whether to estimate a model without covariates.}

\item{subset_outcome}{numeric vector.
Specify position of outcome variables you want to use.}

\item{\dots}{other arguments passed in \code{estimatr::lm_robust}}
}
\value{
list with RCT_OLS class. \code{.$model} is a list of fitted models.
\code{.$res} is a list of result.
}
\description{
This function generates multiple regression models
by passing multiple baseline models
(two-sided formula consisting of outcomes and treatments) and
multiple covariate models (ons-sided formula) as arguments.
In addition, the generated regression model is fitted
by the least squares method of the \code{lm_robust} function
in the {estimatr} package.
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d1 <- sample(c("A", "B", "C"), size = n, replace = TRUE)
d2 <- ifelse(d1 == "A", "control", "treat")
ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- 0.2 * x1 + 0.5 * x2
y1 <- ifelse(d1 == "A", ya, ifelse(d1 == "B", yb, yc))
y2 <- ifelse(y1 > 1, 1, 0)
dt <- data.frame(
  outcome = y1,
  bin_outcome = y2,
  treat = d1,
  bin_treat = d2,
  x1,
  x2
)

# Linear regression
set_optRCTtool(outcome ~ treat, ~ x1 + x2, dt, "A")
rct_lm(data = dt)
rct_lm(ctrl = "B", data = dt)
rct_lm(arms = "bin_treat", ctrl = "control", data = dt)
rct_lm("bin_outcome", data = dt, se_type = "HC1")
rct_lm(xmod = ~x1, data = dt)
set_optRCTtool(outcome + bin_outcome ~ treat, data = dt, ctrl = "A")
rct_lm(data = dt)
rct_lm(data = dt, subset_outcome = 1)
set_optRCTtool(outcome ~ treat, list(~x1, ~x2, ~ x1 + x2), dt, "A")
est <- rct_lm(data = dt)
summary(est)
clear_optRCTtool()


}
