% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_plot.r
\name{rct_plot}
\alias{rct_plot}
\title{Visualization of Empirical Analysis of RCTs}
\usage{
rct_plot(...)
}
\arguments{
\item{\dots}{Specify the arguments to pass to this function.
The first argument must be a dataframe
with a class supported by rct_plot (t_test).
You can check the arguments that can be passed in
the t_test class with \code{help(rct_plot.t_test)}.}
}
\description{
This is an S3 generic function for visualizing RCT analysis.
RCTtoolbox provides a method for t_test (generated by the ttest function).
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))

dt <- data.frame(y, d, x1, x2)

# plot of result of t-test
ttest(y ~ d, dt) \%>\%
  rct_plot(
    label = "{{mean}}[{{effect}}{{star}}]",
    title = "Simulated Outcome"
  )

}
