% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rct.r
\name{rct_plot.t_test}
\alias{rct_plot.t_test}
\title{Visualization of T-test Result}
\usage{
\method{rct_plot}{t_test}(
  data,
  confint = FALSE,
  label = NULL,
  label_digits = 3,
  xlab = "Treatments",
  ylab = NULL,
  title = NULL,
  caption = NULL,
  flip = FALSE,
  ...
)
}
\description{
Visualization of T-test Result
}
