% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_table.r
\name{rct_table.balance_test}
\alias{rct_table.balance_test}
\title{Output Table for Balance Test}
\usage{
\method{rct_table}{balance_test}(
  data,
  digits = 3,
  title = NULL,
  output = "kableExtra",
  footnote = NULL,
  size = 15,
  ...
)
}
\arguments{
\item{data}{data.frame with class "balance_test"}

\item{digits}{numeric.
Specify the number of decimal places to display (Default is 3).}

\item{title}{character. Table title.}

\item{output}{character. Output format.}

\item{footnote}{character. Footnote.}

\item{size}{numeric. Font size.}

\item{\dots}{Other arguments to pass to \code{kableExtra::kable_styling}}
}
\description{
This function outputs
the result of balance test in tabular form.
For tabular output,
this function uses the \code{datasummary} function of the {modelsummary} package
that corresponds to various outputs.
\code{rct_table.power_analysis} supports "kableExtra" (for PDF and HTML output),
"flextable" (MS Word and MS Powerpoint), and "data.frame".
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)
ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))
dt <- data.frame(y, d, x1, x2)

# Balance test with two or more covariates
btest <- balance_test(x1 + x2 ~ d, dt)

# Output table
library(modelsummary)
rct_table(btest, title = "Balance Test")

}
