% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_table.r
\name{rct_table.power_analysis}
\alias{rct_table.power_analysis}
\title{Output Table for Power Analysis}
\usage{
\method{rct_table}{power_analysis}(
  data,
  target = "unstd_d",
  digits = 3,
  title = NULL,
  output,
  footnote,
  size = 13
)
}
\description{
Output Table for Power Analysis
}
