% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_table.r
\name{rct_table.RCT_OLS}
\alias{rct_table.RCT_OLS}
\title{Output Table for Regression Analaysis}
\usage{
\method{rct_table}{RCT_OLS}(
  object,
  coef_map,
  outcome_map = NULL,
  not_show_x = NULL,
  keep_gof = "Num.Obs.|R2|R2 Adj.",
  digits = 3,
  stars = c(`***` = 0.01, `**` = 0.05, `*` = 0.1),
  title = NULL,
  footnote = NULL,
  output = "kableExtra",
  size = 15,
  ...
)
}
\arguments{
\item{object}{object with RCT_OLS class.}

\item{coef_map}{character vector with name.
Specify \code{c("Original variable" = "label", ...)}.}

\item{outcome_map}{character vector with name.
Specify \code{c("Original variable" = "label", ...)}.}

\item{not_show_x}{list.
You can specify how to display information about
whether you are controlling a variable that is not displayed by a factor.
For example, \code{list("label 1" = "x1")} changes the label of "x1" to "label 1".
Also, when specified as \code{list("label" = c ("x1", "x2"))},
if x1 and x2 are not controlled or
the coefficients are not displayed in all models,
"x1" and "x2" labels are aggregated into one line with the name "label".}

\item{keep_gof}{character.
Which regression statistics (e.g. number of observations and R-squares)
to keep?}

\item{digits}{numeric.
Specify the number of decimal places to display (Default is 3).}

\item{stars}{vector. Indicator of p-value.
If \code{c("**" = 0.05, "***" = 0.01)},
then return \verb{***} if p-value <= 0.01,
and return \verb{**} if 0.01 < p-value <= 0.05.
If \code{stars = FALSE}, no stars are output.}

\item{title}{character. Title of table.}

\item{footnote}{character. Table footnote.}

\item{output}{character. Output format.}

\item{size}{numeric. Font size (default is 15).}

\item{\dots}{other arguments passing in kableExtra::kable_styling.}
}
\description{
Output Table for Regression Analaysis
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))

dt <- data.frame(y, d, x1, x2)

# Run regression
est <- rct_lm(y ~ d, xmod = list(~ x1 + x2), data = dt)
rct_table(
  est,
  outcome_map = c("y" = "outcome"),
  coef_map = c("dB" = "State B", "dC" = "State C"),
  not_show_x = list("Covariates" = c("x1", "x2"))
)

}
