% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_table.r
\name{rct_table}
\alias{rct_table}
\title{Output Table of Empirical Analysis for RCTs}
\usage{
rct_table(...)
}
\arguments{
\item{\dots}{Specify the arguments to pass to this function.
The first argument must be a dataframe
with a class supported by rct_table (power_analysis).
You can check the arguments that can be passed in
the power_analysis class by \code{help(rct_table.power_analysis)},
in the balance_test class by \code{help(rct_table.balance_test)},
in the RCT_OLS class by \code{help(rct_table.rct_lm)}.}
}
\description{
This is an S3 generic function
for creating table of RCT analysis.
RCTtoolbox provides a method for power_analysis
(generated by the power_analysis function),
balance_test
(generated by the balance_test function), and
RCT_OLS
(generated by the rct_lm function).
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))
dt <- data.frame(y, d, x1, x2)

# power analysis with variance assumption
tmp <- power_analysis(~d, dt, alpha = 0.05, power = 0.8, std_dev = 0.2)

# Balance test with two or more covariates
btest <- balance_test(~d, c("x1", "x2"), dt)

# table output of power analysis
library(modelsummary)
rct_table(
  tmp,
  title = paste(
    "Power Analysis:",
    "Least Mean Difference to Keep 80\% Power and 5\% Significant Level"
  ),
  footnote = paste(
    "Note: To calculate mean difference,",
    "we assume that standard deviation of outcome is 0.2."
  )
)

# Output table of balance test
rct_table(btest, title = "Balance Test")

# Run regression
est <- rct_lm(y ~ d, xmod = list(~ x1 + x2), data = dt)
rct_table(
  est,
  outcome_map = c("y" = "outcome"),
  coef_map = c("dB" = "State B", "dC" = "State C"),
  not_show_x = list("Covariates" = c("x1", "x2"))
)

}
