% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_plot.r
\name{rctplot}
\alias{rctplot}
\alias{rctplot.RCTtoolbox.ttest}
\title{Visualization of RCTtoolbox}
\usage{
rctplot(...)

\method{rctplot}{RCTtoolbox.ttest}(
  obj,
  zval = 1,
  xlab = "Experimental Arms",
  ylab = "Average (+/- Std.Err.)",
  title = NULL,
  caption = NULL,
  inplot_lab_format = "{round(mean1, 3)}{pstar}",
  inplot_lab_size = 5,
  inplot_lab_adjust = 0,
  ylim = NULL,
  flip = FALSE,
  patchwork = TRUE,
  ncol = NULL,
  nrow = NULL,
  family = getOption("RCTtoolbox.plot_family"),
  ...
)
}
\arguments{
\item{\dots}{arguments of \code{\link[=simplegg]{simplegg()}}}

\item{obj}{R6 object class generated by methods of \code{RCTtoolbox}}

\item{zval}{numeric. z-value to construct error bars
Default is 1 (average +/- standard error).}

\item{xlab}{character. label of x-axis.}

\item{ylab}{character. label of y-axis.}

\item{title}{character. plot title.}

\item{caption}{character. plot caption.}

\item{inplot_lab_format}{character. format label showing results in plot.
Format is based on \code{glue::glue()}.
See \code{help(glue)} in details.}

\item{inplot_lab_size}{numeric. font size of in-plot label.}

\item{inplot_lab_adjust}{numeric. adjust in-plot label on y-axis.}

\item{ylim}{numeric vector. limit of y-axis.}

\item{flip}{logical. whether to swap axises?}

\item{patchwork}{logical.
whether to add a list of plots into one composition?
(Use \code{patchwork::wrap_plots}?)}

\item{ncol}{numeric. dimensions of the grid to create.}

\item{nrow}{numeric. dimensions of the grid to create.}

\item{family}{string of font family.}
}
\description{
Visualization of RCTtoolbox
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- create_RCTtoolbox(
  atest + avacc ~ treat,
  data = subset(RubellaNudge, coupon == 1),
  treat_levels = LETTERS[1:7],
  treat_labels = paste("Message", LETTERS[1:7])
)

rct$ttest()$
  plot(
    title = switch(outcome,
      "atest" = "A. Antibody Test",
      "avacc" = "B. Vaccination"
    ),
    inplot_lab_format = "{round(mean1, 3)}[{round(diff,3)}{pstar}]",
    inplot_lab_adjust = 0.025,
    ylim = c(-0.001, .2),
    flip = TRUE,
    ncol = 1
)
}
}
