% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_table.r
\name{rcttable.balance_test}
\alias{rcttable.balance_test}
\title{Output Table for Balance Test}
\usage{
\method{rcttable}{balance_test}(
  data,
  digits = 3,
  title = NULL,
  output = "kableExtra",
  footnote = NULL,
  size = getOption("RCTtool.table_fontsize"),
  ...
)
}
\arguments{
\item{data}{data.frame with class "balance_test"}

\item{digits}{numeric.
Specify the number of decimal places to display (Default is 3).}

\item{title}{character. Table title.}

\item{output}{character. Output format.}

\item{footnote}{character. Footnote.}

\item{size}{numeric. Font size.}

\item{\dots}{Other arguments to pass to \code{kableExtra::kable_styling}}
}
\description{
This function outputs
the result of balance test in tabular form.
For tabular output,
this function uses the \code{datasummary} function of the {modelsummary} package
that corresponds to various outputs.
\code{rcttable.power_analysis} supports "kableExtra" (for PDF and HTML output),
"flextable" (MS Word and MS Powerpoint), and "data.frame".
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d1 <- sample(c("A", "B", "C"), size = n, replace = TRUE)
d2 <- ifelse(d1 == "A", "control", "treat")
ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d1 == "A", ya, ifelse(d1 == "B", yb, yc))
dt <- data.frame(outcome = y, treat = d1, bin_treat = d2, x1, x2)

# F-test
set_optRCTtool(xmod = ~ x1 + x2, data = dt)
btest <- balance_test(data = dt)

# Output table
library(modelsummary)
rcttable(btest, title = "Balance Test")

}
