% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_table.r
\name{rcttable.RCTtoolbox.balance.test}
\alias{rcttable.RCTtoolbox.balance.test}
\title{Output Table for Balance Test}
\usage{
\method{rcttable}{RCTtoolobox.balance.test}(
  obj,
  treat.label = "Treatments",
  digits = 3,
  title = NULL,
  footnote = NULL,
  size = 12,
  output = "kableExtra",
  ...
)
}
\arguments{
\item{digits}{numeric.
Specify the number of decimal places to display (Default is 3).}

\item{title}{character. Table title.}

\item{footnote}{character. Footnote.}

\item{size}{numeric. Font size.}

\item{output}{character. Output format.}

\item{\dots}{Other arguments to pass to \code{kableExtra::kable_styling}}

\item{data}{data.frame with class "balance_test"}
}
\description{
This function outputs
the result of balance test in tabular form.
For tabular output,
this function uses the \code{datasummary} function of the {modelsummary} package
that corresponds to various outputs.
\code{rcttable.power_analysis} supports "kableExtra" (for PDF and HTML output),
"flextable" (MS Word and MS Powerpoint), and "data.frame".
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- create_RCTtoolbox(
  atest + avacc ~ treat,
  ~ age + educ,
  RubellaNudge,
  LETTERS[1:7]
)

rct$balance(subset = coupon == 1)$table()

}

}
