% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_table.r
\name{rcttable.RCTtoolbox.lm}
\alias{rcttable.RCTtoolbox.lm}
\title{Output Table for Regression Analaysis}
\usage{
\method{rcttable}{RCTtoolbox.lm}(
  object,
  dvar,
  add_coef_map = NULL,
  outcome_map = NULL,
  not_show_x = NULL,
  keep_gof = "Num.Obs.|R2|R2 Adj.",
  digits = 3,
  stars = c(`***` = 0.01, `**` = 0.05, `*` = 0.1),
  title = NULL,
  footnote = NULL,
  size = getOption("RCTtoolbox.table_font_size"),
  output = getOption("RCTtoolbox.table_output"),
  ...
)
}
\arguments{
\item{object}{object with R6 class "RCTtoolbox.lm"}

\item{dvar}{string of variable name of experimental arms.}

\item{add_coef_map}{character vector with name.
Specify \code{c("Original variable" = "label", ...)}.}

\item{outcome_map}{character vector with name.
Specify \code{c("Original variable" = "label", ...)}.}

\item{not_show_x}{list.
You can specify how to display information about
whether you are controlling a variable that is not displayed by a factor.
For example, \code{list("label 1" = "x1")} changes the label of "x1" to "label 1".
Also, when specified as \code{list("label" = c ("x1", "x2"))},
if x1 and x2 are not controlled or
the coefficients are not displayed in all models,
"x1" and "x2" labels are aggregated into one line with the name "label".}

\item{keep_gof}{character.
Which regression statistics (e.g. number of observations and R-squares)
to keep?}

\item{digits}{numeric.
Specify the number of decimal places to display (Default is 3).}

\item{stars}{vector. Indicator of p-value.
If \code{c("**" = 0.05, "***" = 0.01)},
then return \verb{***} if p-value <= 0.01,
and return \verb{**} if 0.01 < p-value <= 0.05.
If \code{stars = FALSE}, no stars are output.}

\item{title}{character. Title of table.}

\item{footnote}{character. Table footnote.}

\item{size}{numeric. Font size (default is 15).}

\item{output}{character. Output format.}

\item{\dots}{other arguments passing in kableExtra::kable_styling.}
}
\description{
Output Table for Regression Analaysis
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- create_RCTtoolbox(
  atest + avacc ~ treat,
  list(~age, ~ age + educ),
  RubellaNudge,
  LETTERS[1:7],
  letters[1:7]
)

rct$
  lm(subset = coupon == 1)$
  est$table(
    add_coef_map = c("age" = "Age"),
    outcome_map = c("atest" = "Test", "avacc" = "Vaccination"),
    not_show_x = list(Covariates = c("educ")),
    digits = 4,
    keep_gof = "Num.Obs.|R2"
  )
}

}
