% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_table.r
\name{rcttable}
\alias{rcttable}
\title{Output Table of Empirical Analysis for RCTs}
\usage{
rcttable(...)
}
\arguments{
\item{\dots}{Specify the arguments to pass to this function.
The first argument must be a dataframe
with a class supported by rcttable (power_analysis).
You can check the arguments that can be passed in
the power_analysis class by \code{help(rcttable.power_analysis)},
in the balance_test class by \code{help(rcttable.balance_test)},
in the RCT_OLS class by \code{help(rcttable.rct_lm)}.}
}
\description{
This is an S3 generic function
for creating table of RCT analysis.
RCTtoolbox provides a method for power_analysis
(generated by the power_analysis function),
balance_test
(generated by the balance_test function), and
RCT_OLS
(generated by the rct_lm function).
}
