% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleaning.r
\name{reorder_arms}
\alias{reorder_arms}
\title{Reorder of treatment arms}
\usage{
reorder_arms(treat_levels = NULL, treat_labels = NULL, ctrl = NULL)
}
\arguments{
\item{treat_levels}{original level of treatment arms.
The first level is control arm.}

\item{treat_labels}{label of treatment arms corresponding to original level.}

\item{ctrl}{new control arm.}
}
\description{
Reorder of treatment arms
}
