% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-func.r
\name{se}
\alias{se}
\title{Calculation Standard Errors of Mean}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector.}

\item{na.rm}{logical.
Specify whether to calculate without missing values.}
}
\description{
Calculation Standard Errors of Mean
}
\examples{
set.seed(120511)
a <- rnorm(100)
mean(a)
se(a)
}
