% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optRCTtool.r
\name{set_optRCTtool}
\alias{set_optRCTtool}
\title{Manipulation Global Options for RCTtoolbox}
\usage{
set_optRCTtool(basicmod, xmod, data, ctrl, ...)
}
\arguments{
\item{basicmod}{two-sided formula with
outcome variable on lhs and treatment variable on rhs}

\item{xmod}{(list of) one-sided formulas with
covariates used in regression analysis and balance test on the rhs.}

\item{data}{data you want to use}

\item{ctrl}{string. Label of the control.
If missing, the first level of factor of treatment variable
is the control.}

\item{\dots}{Specify \code{option_name = option}}
}
\description{
Manipulation Global Options for RCTtoolbox
}
\examples{
\dontrun{
set.seed(120511)
ex <- data.frame(treat = sample(c("A", "B"), size = 100, replace = TRUE))
ex$outcome <- ifelse(ex$treat == "A", rnorm(1, mean = 2), rnorm(1))

set_optRCTtool(
  outcome ~ treat,
  data = ex,
  ctrl = "B",
  RCTtool.table_fontsize = 8
)
}

}
