% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRCTtool.r
\name{setRCTtool}
\alias{setRCTtool}
\title{Set global options for RCTtoolbox}
\usage{
setRCTtool(...)
}
\arguments{
\item{\dots}{Specify \code{option_name = option}}
}
\description{
Set global options for RCTtoolbox
}
\examples{
\dontrun{
setRCTtool(
  RCTtool.outcome = c("y1", "y2"),
  RCTtool.treatment = ~ d,
  RCTtool.nothing = "" # not register by .onLoad
)
getOption("RCTtool.outcome")
getOption("RCTtool.treatment")
getOption("RCTtool.nothing") # not register by .onLoad
}

}
