% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_lm.r
\name{summary.RCT_OLS}
\alias{summary.RCT_OLS}
\title{Summary Method for RCT_OLS class}
\usage{
\method{summary}{RCT_OLS}(object, ...)
}
\arguments{
\item{object}{object RCT_OLS class}

\item{\dots}{other arguments passing in summary function.}
}
\description{
Summary Method for RCT_OLS class
}
