% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.r
\name{summary.RCTtoolbox.lm}
\alias{summary.RCTtoolbox.lm}
\title{Summary of linear model}
\usage{
\method{summary}{RCTtoolbox.lm}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary of linear model
}
