% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.r
\name{summary.RCTtoolbox.ttest}
\alias{summary.RCTtoolbox.ttest}
\title{Summary of t-test}
\usage{
\method{summary}{RCTtoolbox.ttest}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary of t-test
}
