% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-func.r
\name{ttest}
\alias{ttest}
\title{T-Test for RCTs with Multiple Experimental Arms}
\usage{
ttest(y, arms, arms_label, ctrl, data)
}
\arguments{
\item{y}{a string of outcome.}

\item{arms}{a string of treatment variables.}

\item{arms_label}{a string vector of label of treatments.}

\item{ctrl}{a string of the control.}

\item{data}{data object you want to use.}
}
\description{
T-Test for RCTs with Multiple Experimental Arms
}
