% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_maker.r
\name{ttest_label_maker}
\alias{ttest_label_maker}
\title{Making label for Visualization}
\usage{
ttest_label_maker(label, mean, se, effect, p, digits = 3)
}
\arguments{
\item{label}{character. Specify label format.}

\item{mean}{numeric. Specify outcome average.}

\item{se}{numeric. Specify standard error of outcome average.}

\item{effect}{numeric. Specify treatment effect.}

\item{p}{numeric. Specify p-value of treatment effect.}

\item{digits}{numeric. Specify how many decimal places to display.
Default is 3.}
}
\description{
Making label for Visualization
}
