% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_ttest.r
\name{ttest_multi_mod_arm}
\alias{ttest_multi_mod_arm}
\title{Method of RCTtoolbox: \code{ttest()}}
\usage{
ttest_multi_mod_arm(
  baseline = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = NULL,
  ctrl = NULL,
  subset = NULL,
  weights = NULL,
  bootse = 0L,
  bootp = 0L,
  seed = 120511L
)
}
\arguments{
\item{baseline}{list of two-sided formula \code{outcome ~ treat}.
The method \code{ttest()} automatically
passes the private field \code{formula.yd} of R6 object "RCTtoolbox"
to this argument.}

\item{data}{data.frame/tibble object that you want to use.
The method \code{ttest()} automatically
passes a public field \code{data} of R6 object "RCTtoolbox"
to this argument.}

\item{treat_levels}{character vector. Level of experimental arms.
The first element is control arm.
The method \code{ttest()} automatically
passes the private field \code{dvec.levels} of R6 object "RCTtoolbox"
to this argument.}

\item{treat_labels}{character vector. Label of experimental arms
corresponding to \code{treat_levels}.
The method \code{ttest()} automatically
passes the private field \code{dvec.labels} of R6 object "RCTtoolbox"
to this argument.}

\item{ctrl}{character vector. New control arm.
If NULL (default),
the first element of \code{treat_levels} is control arm.
The method \code{ttest()} has this argument.}

\item{subset}{subset condition.
If NULL (default),
full observations of \code{data} is used to implement t-test.
The method \code{ttest()} has this argument.}

\item{weights}{weight variable.
If NULL (default),
unbiased unweighted average and std.err. is calculated.
The method \code{ttest()} has this argument.}

\item{bootse}{integer.
Number of bootstrap samples to calculate std.err. of mean.
If 0 (default),
theoretical std.err. of mean is calculated.
The method \code{ttest()} has this argument.}

\item{bootp}{integer.
If 0 (default), welch two-sided t-test is implemented.
If 1 or more, the permutation test is implemented
(the number of permutations that randomly reallocate assignment
to keep the sample size collected from each group).
The method \code{ttest()} has this argument.}

\item{seed}{integer.
Seed value passed to the argument of \code{\link{set.seed}}.
The method \code{ttest()} has this argument.}
}
\value{
data frame with 13 variables
\describe{
\item{mean1}{Outcome average of each experimental arm}
\item{se1}{Std.err. of outcome average of each experimental arm}
\item{n1}{Number of observations of each experimental arm}
\item{mean0}{Outcome average of control arm}
\item{se0}{Std.err. of outcome average of control arm}
\item{n0}{Number of observations of control arm}
\item{diff}{Difference between "mean1" - "mean0"}
\item{t}{t-statistics when \code{bootp = 0L} (Two-sided t-test)}
\item{df}{degree of freedom of t-distribution
when \code{bootp = 0L} (Two-sided t-test)}
\item{pval}{p-value of difference mean test.
Pr(>|t|) when \code{bootp = 0L} (Two-sided t-test)
Pr(>|diff|) when \code{bootp > 0L} (Permutation test)}
\item{method}{Character. Method of difference mean test.}
\item{arms}{Factor. Experimental arms}
\item{outcome}{Character. Outcome variable.}
}
}
\description{
When you create R6 object "RCTtoolbox"
via \code{\link{create_RCTtoolbox}}
and run its method \code{ttest()},
the method internally implements this function
which performs two-sided t-test or permutation test.
}
\details{
A method \code{ttest()} internally run the function
\code{function(...) {ttest_multi_mod_arm(
private$formula.yd, self$data,
private$dvec.levels, private$dvec.labels, ...
)}}.
Thus, you can provide additional arguments
\code{ctrl}, \code{subset}, \code{weights}, \code{bootse},
\code{bootp}, and \code{seed} via dot-dot-dot.
}
\section{Return of method \code{ttest()}}{


When you run \code{ttest()} which is a method of RCTtoolbox class,
it internally implements \code{ttest_multi_mod_arm}
and create new R6 object with "RCTtoolbox.ttest" class.
The returned object of \code{ttest_multi_mod_arm}
is saved in a field \code{result}.
Thus, you can access it via \code{$result}.
RCTtoolbox.ttest class has two methods related to output:
\describe{
\item{\code{plot()}}{Visualization of result.}
\item{\code{summary()}}{Print result in console.}
}
}

\examples{
\dontrun{
  data(RubellaNudge)
  rct <- create_RCTtoolbox(
    atest + avacc ~ treat,
    data = RubellaNudge,
    treat_levels = LETTERS[1:7]
  )

  # two-sided t-test
  rct$ttest()
  # permutation test with control arm "C"
  rct$ttest(bootp = 50, ctrl = "C")
}

}
