% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_power.r
\name{ttest_power}
\alias{ttest_power}
\title{Power Analysis for Two Sided T-Test}
\usage{
ttest_power(n0, n1, d, alpha, power)
}
\arguments{
\item{n0}{numeric. Number of observations in the control.}

\item{n1}{numeric. Number of observations in the treatment.}

\item{d}{numeric. Effect size.}

\item{alpha}{numeric. Statistically significance
(probability of Type I error).}

\item{power}{numeric. Power (1 - probability of Type II error).}
}
\description{
This is a function that performs
a power analysis of the two-sided t-test.
Given the effect size, significance level, and power,
the required sample size can be calculated
(although the control or treatment sample size must be given).
Also, given the sample size of the two groups, effect size,
and significance level, the power can be calculated.
Given the sample size of two groups, significance level,
and power of the two groups, the effect size can be calculated.
Given the sample size of two groups, effect size,
and power of the two groups, the significance level can be calculated.
}
