% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_ttest.r
\name{wtd_mean}
\alias{wtd_mean}
\title{Calculate sample average}
\usage{
wtd_mean(x, w = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{w}{numeric vector of weight.
If NULL, unweighted sample average is calculated.}
}
\description{
Calculate sample average
}
