% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_ttest.r
\name{wtd_ttest}
\alias{wtd_ttest}
\title{Perform Two-Sided T-Test/Permutation Test}
\usage{
wtd_ttest(
  y1,
  y0,
  w1 = NULL,
  w0 = NULL,
  bootse = NULL,
  bootp = NULL,
  seed = 120511
)
}
\arguments{
\item{y1}{numeric. Outcome in group 1.}

\item{y0}{numeric. Outcome in group 0.}

\item{w1}{numeric. Weight in group 1.
If NULL, unweighted sample average is calculated.}

\item{w0}{numeric. Weight in group 0.
If NULL, unweighted sample average is calculated.}

\item{bootse}{numeric. Number of bootstrap sample to calculate se.
If NULL, se is calculated by effective sample size.}

\item{bootp}{numeric. Number of re-randomization
to conduct permutation test.
If NULL, perform (welch) two-sided t-test.}

\item{seed}{numeric. seed value.}
}
\description{
Perform Two-Sided T-Test/Permutation Test
}
