% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_ttest.r
\name{wtd_var}
\alias{wtd_var}
\title{Calculate unbiased variance and standard error}
\usage{
wtd_var(x, w = NULL, se = TRUE, bootse = NULL, seed = 120511)
}
\arguments{
\item{x}{numeric vector}

\item{w}{numeric vector of weight.
If NULL, unweighted unbiased estimator is calculated.}

\item{se}{logical. Return standard error (TRUE) or variance (FALSE)}

\item{bootse}{numeric. Number of bootstrap sample to calculate se.}

\item{seed}{numeric. seed value.}
}
\description{
Calculate unbiased variance and standard error
}
