% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_test.r
\name{balance_test}
\alias{balance_test}
\title{Blance Test by F-test of Overall Siginificance in Linear Regression}
\usage{
balance_test(mod, data)
}
\arguments{
\item{mod}{formula such as \code{covariate1 + covariate2 ~ treatment}.
If missing, try to find out experimental arms and covariates
from global options "RCTtool.arms" and "RCTtool.xmod"
If formula with rhs only,
try to find out covariates from global options "RCTtool.xmod".
If formula with lhs only,
try to find out experimental arms
from global options "RCTtool.arms"}

\item{data}{data which you want to use.}
}
\value{
A data frame (`balance_test`` class) with 3 variables:
\describe{
\item{x}{character of covariates used for balance test}
\item{item}{character of treatment arms and F-test}
\item{val}{outcome mean of each arm and p-value of F-test}
}
}
\description{
This function performs a balance test of the treatment
with multiple covariates.
The balance test calculates the mean of each group
and the F-test by linear regression analysis.
User specifies the treatment variable with a one-sided formula
and the covariate used for the balance test with a string vector.
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))
dt <- data.frame(y, d, x1, x2)

# use mod argument
balance_test(x1 ~ d, data = dt)
balance_test(x1 + x2 ~ d, data = dt)

# use global options
optRCTtool(
  RCTtool.arms = ~d,
  RCTtool.xmod = list(~x1, ~x2, ~ x1 + x2)
)
balance_test(data = dt)
balance_test(x1 ~ 0, data = dt)
optRCTtool(clear = TRUE)


}
