% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_balance.r
\name{balance_test}
\alias{balance_test}
\title{F-test of Overall Siginificance in Linear Regression}
\usage{
balance_test(
  baseline = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = NULL,
  ctrl = NULL,
  subset = NULL,
  weights = NULL,
  cluster = NULL,
  ...
)
}
\arguments{
\item{baseline}{two-sided formula \code{covariate ~ treat}}

\item{data}{data.frame/tibble you want to use}

\item{treat_levels}{order of experimental arms}

\item{ctrl}{specify control arm if you want change}

\item{subset}{subset condition of data}

\item{weights}{weight variable}

\item{cluster}{cluster variable}

\item{\dots}{pass to \code{estimatr::lm_robust()}}
}
\description{
F-test of Overall Siginificance in Linear Regression
}
