% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_test.r
\name{balance_test_multi_var}
\alias{balance_test_multi_var}
\title{Blance Test by F-test of Overall Significance in Linear Regression}
\usage{
balance_test_multi_var(
  covariate = NULL,
  treat = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = NULL,
  ctrl = NULL,
  subset = NULL,
  weights = NULL,
  cluster = NULL,
  ...
)
}
\arguments{
\item{data}{data which you want to use.}

\item{xlist}{string vector with covariates.
If missing, try to find form \code{option("RCTtool.xlist")}}

\item{arms}{string of treatment variables.
If missing, try to find from \code{option("RCTtool.arms")}}
}
\value{
A data frame with \code{balance_test} class
}
\description{
This function performs a balance test of the treatment
with multiple covariates.
The balance test calculates the mean of each group
and the F-test by linear regression analysis.
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- create_RCTtoolbox(
  atest + avacc ~ treat,
  ~ age + educ,
  RubellaNudge,
  LETTERS[1:7]
)

rct$balance(subset = coupon == 1)$result

}


}
