% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_RCTdata.r
\name{clean_RCTdata}
\alias{clean_RCTdata}
\title{Shape dataset}
\usage{
clean_RCTdata(
  baseline,
  covariate = NULL,
  data,
  subset = NULL,
  weights = NULL,
  cluster = NULL
)
}
\arguments{
\item{baseline}{formula. \code{outcome ~ treatment}.}

\item{covariate}{one-sided formula. \code{. ~ . + x1 + x2}.}

\item{data}{data.frame/tibble object you want to use.}

\item{subset}{subset condition.}

\item{weights}{weight variable.}

\item{cluster}{cluster variable.}
}
\description{
Shape dataset
}
