% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_create.r
\name{create_RCTtoolbox}
\alias{create_RCTtoolbox}
\title{Create New Toolbox}
\usage{
create_RCTtoolbox(
  baseline = NULL,
  covariate = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = treat_levels
)
}
\arguments{
\item{baseline}{two-sided formula with
one or more outcome variables on LHS and
treatment variable on RHS.}

\item{covariate}{(list of) one-sided formulas with
covariates used in regression and balance test on RHS.}

\item{data}{data.frame/tibble object you want to use.}

\item{treat_levels}{character vector. Levels of experimental arms.
The first element is control arm.}

\item{treat_labels}{character vector.
Labels of experimental arms corresponding to \code{treat_levels}.
Default is to use \code{treat_levels} as labels.}
}
\value{
R6 object with "RCTtoolbox" class.
The returned object has following fields and methods
\itemize{
\item \code{data}: Field.
Store data.frame/tibble passed to \code{data} argument.
You can access via \code{$data}
\item \code{print()}: Method.
Print information about the returned object.
Run \code{$print()}
\item \code{ttest()}: Method.
Implement two-sided t-test or permutation test.
Run \code{$ttest()}
See \code{\link{ttest_multi_mod_arm}} for help.
\item \code{power()}: Method.
Implement power analysis.
Run \code{$power()}
\item \code{balance()}: Method.
Implement balance test.
Run \code{$balance()}
\item \code{lm()}: Method.
Estimate linear model.
Run \code{$lm()}
}
}
\description{
Create an R6 class object called "RCTtoolbox".
}
\examples{
\dontrun{
data(RubellaNudge)
create_RCTtoolbox(
  baseline = itest + ivacc ~ treat,
  covariate = ~ age + educ,
  data = RubellaNudge,
  treat_levels = LETTERS[1:7]
)
}


}
