% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_RCTtool.r
\name{create_RCTtoolbox}
\alias{create_RCTtoolbox}
\title{Create New Toolbox}
\usage{
create_RCTtoolbox(baseline, covariate = NULL, data, ctrl, direct = NULL, ...)
}
\arguments{
\item{baseline}{two-sided formula with
outcome variables on LHS and treatment variable on RHS.}

\item{covariate}{(list of) one-sided formulas with
covariates used in regression and balance test on RHS.}

\item{data}{data.frame/tibble object you want to use.}

\item{ctrl}{string of the control arm.}

\item{direct}{string of direct calling.}

\item{\dots}{other arguments}
}
\description{
Create an R6 class object called "RCT toolbox"
for consistent RCT data analysis.
The created object has method functions
such as t-test and regression analysis.
}
\examples{
\dontrun{
data(RubellaNudge)
create_RCTtoolbox(
  baseline = itest + ivacc ~ treat,
  covariate = ~ age + educ,
  data = RubellaNudge,
  ctrl = "A"
)
}


}
