% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_create.r
\name{create_RCTtoolbox}
\alias{create_RCTtoolbox}
\title{Create New Toolbox}
\usage{
create_RCTtoolbox(baseline, covariate = NULL, data, levels, labels = NULL, ...)
}
\arguments{
\item{baseline}{two-sided formula with
outcome variables on LHS and treatment variable on RHS.}

\item{covariate}{(list of) one-sided formulas with
covariates used in regression and balance test on RHS.}

\item{data}{data.frame/tibble object you want to use.}

\item{levels}{character vector of levels of experimental arms.
The first element must be control.}

\item{labels}{character vector of labels of experimental arms.}

\item{\dots}{other arguments}
}
\description{
Create an R6 class object called "RCT toolbox"
for consistent RCT data analysis.
The created object has method functions
such as t-test and regression analysis.
}
\examples{
\dontrun{
data(RubellaNudge)
create_RCTtoolbox(
  baseline = itest + ivacc ~ treat,
  covariate = ~ age + educ,
  data = RubellaNudge,
  ctrl = "A"
)
}


}
