% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorlize.r
\name{factorlize}
\alias{factorlize}
\title{Convert to Factor Variables}
\usage{
factorlize(x, base)
}
\arguments{
\item{x}{vector}

\item{base}{character. Name of the control.}
}
\description{
Convert to Factor Variables
}
