% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-func.r
\name{ftest}
\alias{ftest}
\title{F-test of Overall Siginificance in Linear Regression}
\usage{
ftest(x, arms, arms_label, data)
}
\arguments{
\item{x}{a string of covariates}

\item{arms}{a string of treatment variables}

\item{arms_label}{a string vector with labels of treatment arms.}

\item{data}{data which you want to use.}
}
\description{
F-test of Overall Siginificance in Linear Regression
}
