% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_power.r
\name{power_calculation}
\alias{power_calculation}
\title{Power Analysis of Two Sided T-Test for RCT with Multiple Arms}
\usage{
power_calculation(
  treat = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = NULL,
  ctrl = NULL,
  subset = NULL,
  sd = 1,
  ...
)
}
\arguments{
\item{treat}{A string of treatment variables}

\item{data}{data which you want to use.}

\item{treat_levels}{original level of treatment arms.
The first level is control arm.}

\item{treat_labels}{label of treatment arms corresponding to original level.}

\item{ctrl}{new control arm.}

\item{subset}{subset condition.}

\item{sd}{numeric (default is 1).
The standard deviation used when calculating the mean difference.}

\item{\dots}{some arguments passing in \code{ttest_power}.
You can pass \code{d} (effect size), \code{alpha} (significant level), and
\code{power} (power).}
}
\description{
This function performs a two-sided t-test power analysis
given the sample sizes of the treatment and control groups.
Calculate the missing value among the effect size, significance level,
and power.
For the effect size, the non-standardized effect
(absolute value of the difference between the average outcomes)
is also calculated.
If some variable is specified in the \code{std_dev} argument,
the square root of the mean of variances of specified variable
of the two group is the standard deviation.
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- create_RCTtoolbox(
  atest + avacc ~ treat,
  data = RubellaNudge,
  treat_levels = LETTERS[1:7]
)

rct$power(alpha = 0.05, power = 0.8, sd = 0.2)$result
}


}
