% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_lm.r
\name{rct_lm}
\alias{rct_lm}
\title{Fitting Models by OLS}
\usage{
rct_lm(
  baseline = NULL,
  covariate = NULL,
  data = NULL,
  treat_levels = NULL,
  treat_labels = NULL,
  ctrl = NULL,
  subset = NULL,
  weights = NULL,
  cluster = NULL,
  only_dmod = TRUE,
  ...
)
}
\arguments{
\item{baseline}{two-sided formula with
outcome variables on LHS and treatment variable on RHS.}

\item{covariate}{(list of) one-sided formulas with
covariates used in regression and balance test on RHS.}

\item{data}{data.frame/tibble object you want to use.}

\item{treat_levels}{character vector of levels of experimental arms.
The first element must be control.}

\item{treat_labels}{character vector of labels of experimental arms.}

\item{ctrl}{string of new control arm}

\item{subset}{subset condition of data}

\item{weights}{weight variable}

\item{cluster}{cluster variable}

\item{only_dmod}{logical. estimate linear model without covariates?}

\item{\dots}{other arguments passed in \code{estimatr::lm_robust}}
}
\description{
This function generates multiple regression models
by passing multiple baseline models
(two-sided formula consisting of outcomes and treatments) and
multiple covariate models (ons-sided formula) as arguments.
In addition, the generated regression model is fitted
by the least squares method of the \code{lm_robust} function
in the {estimatr} package.
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- create_RCTtoolbox(
  atest + avacc ~ treat,
  ~ age + educ,
  RubellaNudge,
  LETTERS[1:7]
)

rct$lm(subset = coupon == 1)$summary()

}


}
