% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_lm.r
\name{rct_lm}
\alias{rct_lm}
\title{Fitting Models by OLS}
\usage{
rct_lm(basemod, xmod = NULL, include_onlyd = TRUE, base = NULL, data, ...)
}
\arguments{
\item{basemod}{list of baseline model.
Baseline model is specified by \code{y ~ d}
where \code{y} is outcome, and \code{d} is treatments.}

\item{xmod}{list of covariate model.
Covariate model is specified by one-sided formula like \code{~ x1 + x2}.}

\item{include_onlyd}{logical.
Whether to estimate a model without covariates.}

\item{base}{character. Name of the control.
This argument is passed when treatment variable is not factor.}

\item{data}{data or list of data.
Length of data list must be same as
length of model list generated by genmod.}

\item{\dots}{other arguments passed in \code{estimatr::lm_robust}}
}
\value{
list with RCT_OLS class. \code{.$model} is a list of fitted models.
\code{.$res} is a list of result.
}
\description{
This function generates multiple regression models
by passing multiple baseline models
(two-sided formula consisting of outcomes and treatments) and
multiple covariate models (ons-sided formula) as arguments.
In addition, the generated regression model is fitted
by the least squares method of the \code{lm_robust} function
in the {estimatr} package.
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))

dt <- data.frame(y, d, x1, x2)

# Run regression
est <- rct_lm(y ~ d, xmod = ~ x1 + x2, data = dt)
summary(est)

}
