% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_plot.r
\name{rct_plot}
\alias{rct_plot}
\title{Visualization of Empirical Analysis of RCTs}
\usage{
rct_plot(...)
}
\arguments{
\item{\dots}{Specify the arguments to pass to this function.
The first argument must be a dataframe
with a class supported by rct_plot (t_test).
You can check the arguments that can be passed in
the t_test class with \code{\link[=rct_plot.t_test]{rct_plot.t_test()}}.}
}
\description{
This is an S3 generic function for visualizing RCT analysis.
RCTtoolbox provides a method for t_test (generated by the ttest function).
}
