% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_plot.r
\name{rct_plot.t_test}
\alias{rct_plot.t_test}
\title{Visualization of T-test Result}
\usage{
\method{rct_plot}{t_test}(
  data,
  confint = FALSE,
  label = NULL,
  label_digits = 3,
  xlab = "Treatments",
  ylab = NULL,
  title = NULL,
  caption = NULL,
  text_size = 5,
  text_adjust = 0.2,
  font_family = NULL,
  flip = FALSE,
  ...
)
}
\arguments{
\item{data}{tibble or data.frame with "t_test" class.}

\item{confint}{logical.
Specify whether error bars represent 95\% confidence intervals.}

\item{label}{character. See details.}

\item{label_digits}{numeric. See details.}

\item{xlab}{character. Specify title of x-axis.
Default is "Treatments".}

\item{ylab}{character. Specify title of y-axis.
Default is NULL.
If NULL, the title of y-axis is "Average (+/- Std.Err)" or
"Average (+/- 95\%CI)".}

\item{title}{character. Specify title of plot.
Default is NULL.
If NULL, the label of outcome variable is plot title.}

\item{caption}{character. Specify caption of plot.}

\item{text_size}{numeric. Sepecify font size of label.
Default is 5.}

\item{text_adjust}{numeric.
Specify a value that corrects the label position along the y-axis.}

\item{font_family}{character. Specify font family. Default is NULL.}

\item{flip}{logical. Specify whether to swap the x-axis and y-axis.
Default is FALSE.}

\item{\dots}{other arugments passing \code{simplegg}.}
}
\description{
This is a function that visualizes
the result of the t-test executed by the \code{ttest} function with {ggplot2}.
}
\details{
By using the \code{label} argument,
the result of the t-test can be embedded in the plot as text.
The result of the t-test can be specified by \{\{type\}\}.
If \{\{mean\}\}, the mean value of each group is embedded.
If \{\{se\}\}, the standard error of the mean of each group is embedded.
If \{\{effect\}\}, the effect of each group is embedded.
If \{\{pvalue\}\}, the p-value of each group effect is embedded.
If \{\{star\}\}, the p-value star of each group effect is embedded.
Combining these four, you specify the output of the text.
For example, if the mean is 0.5 and its standard error is 0.1,
then specifying "\{\{mean\}\}(\{\{se\}\})" will embed "0.5(0.1)" in the plot.
The font size of label can by specified by \code{text_size} argument.
The default is 5.
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))

dt <- data.frame(y, d, x1, x2)

# plot of result of t-test
ttest(y ~ d, dt) \%>\%
  rct_plot(
    label = "{{mean}} [{{effect}}{{star}}]",
    title = "Simulated Outcome"
  )

}
