% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_table.r
\name{rct_table.power_analysis}
\alias{rct_table.power_analysis}
\title{Output Table for Power Analysis}
\usage{
\method{rct_table}{power_analysis}(
  data,
  tar = "diff_mean",
  dlab = "Treatments",
  tardigits = 3,
  tarlab = "Mean Difference",
  title = NULL,
  footnote = NULL,
  size = getOption("RCTtool.table_fontsize"),
  output = "kableExtra",
  ...
)
}
\arguments{
\item{data}{dataframe with class "power_analysis"}

\item{tar}{character. Specify which value to output.
If "d", the effect size is output.
If "diff_mean" (default), the non-standarized effect size
(i.e., absolute value of mean difference) is output.
If "alpha", the significant level is output.
If "power", the power is output.}

\item{dlab}{character. Label of treatment column.}

\item{tardigits}{numeric.
Specify the number of decimal places to display (Default is 3).}

\item{tarlab}{character. Label of output \code{tar} column.}

\item{title}{character. Table title.}

\item{footnote}{character. Footnote (kableExtra or flextable).}

\item{size}{numeric. Font size (kableExtra or flextable).}

\item{output}{character. Output format.}

\item{\dots}{Other arguments to pass to \code{kableExtra::kable_styling}}
}
\description{
This function outputs
the result of power analysis in tabular form.
For tabular output,
this function uses the \code{datasummary} function of the {modelsummary} package
that corresponds to various outputs.
\code{rct_table.power_analysis} supports "kableExtra" (for PDF and HTML output),
"flextable" (MS Word and MS Powerpoint), and "data.frame".
}
\examples{
set.seed(120511)
d1 <- sample(LETTERS[1:5], size = 1000, replace = TRUE)
d2 <- ifelse(d1 \%in\% LETTERS[1:3], "Control", "Treat")
ex <- data.frame(treat = d1, treat2 = d2)
set_optRCTtool(~treat, data = ex, ctrl = "A")
est <- power_analysis(data = ex, alpha = 0.05, power = 0.8)

# table output of power analysis
library(modelsummary)
rct_table(
  est,
  dlab = "Experimental arms",
  title = paste(
    "Power Analysis:",
    "Least Mean Difference to Keep 80\% Power and 5\% Significant Level"
  ),
  footnote = paste(
    "Note: To calculate mean difference,",
    "we assume that standard deviation of outcome is 0.2."
  )
)

}
