% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_table.r
\name{rct_table.RCT_OLS}
\alias{rct_table.RCT_OLS}
\title{Output Table for Regression Analaysis}
\usage{
\method{rct_table}{RCT_OLS}(
  object,
  coefmap,
  noshow_x,
  gofkeep = "Num.Obs.|R2|R2 Adj.",
  digits = 3,
  title,
  footnote,
  output = "kableExtra",
  size = 15
)
}
\description{
Output Table for Regression Analaysis
}
