% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_table.r
\name{rct_table}
\alias{rct_table}
\title{Output Table of Empirical Analysis for RCTs}
\usage{
rct_table(...)
}
\arguments{
\item{\dots}{Specify the arguments to pass to this function.
The first argument must be a dataframe
with a class supported by rct_table (power_analysis).
You can check the arguments that can be passed in
the power_analysis class by \code{help(rct_table.power_analysis)},
in the balance_test class by \code{help(rct_table.balance_test)},
in the RCT_OLS class by \code{help(rct_table.rct_lm)}.}
}
\description{
This is an S3 generic function
for creating table of RCT analysis.
RCTtoolbox provides a method for power_analysis
(generated by the power_analysis function),
balance_test
(generated by the balance_test function), and
RCT_OLS
(generated by the rct_lm function).
}
