% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{RCTtoolbox}
\alias{RCTtoolbox}
\alias{RCTtoolbox-package}
\title{Useful Wrapper Functions for RCTs}
\description{
This package is a collection of useful wrapper functions for analyzing RCT data.
    It corresponds to three or more treatment conditions.
    It also offers global options to facilitate consistent analysis.
}
\section{Package Options}{


RCTtoolbox provides the following \code{\link[=options]{options()}}
to facilitate consistent analysis:
\itemize{
\item \code{RCTtoolbox.table_output}: Table output package
("kableExtra", "flextable")
\item \code{RCTtoolbox.table_fontsize}: The font size in the table
output by {kableExtra} and {flextable} used through {modelsummary}.
\item \code{RCTtoolbox.plot_family}: The name of the font family
used to output the figure with {ggplot2}.
}
}

\author{
\strong{Maintainer}: Hiroki Kato \email{vge008kh@student.econ.osaka-u.ac.jp}

}
