% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{RCTtoolbox}
\alias{RCTtoolbox}
\alias{RCTtoolbox-package}
\title{Useful Wrapper Functions for RCTs}
\description{
What the package does (one paragraph).
}
\section{Package Options}{


RCTtoolbox provides the following \code{\link[=options]{options()}}
to facilitate consistent analysis:
\itemize{
\item \code{RCTtool.outcome}: One-sided formula with
outcome variables on the left-hand side of the formula
(e.g. \code{y1 + y2 ~ 0})
\item \code{RCTtool.arms}: One-sided formula with
treatment variables on the right-hand side of the formula
(e.g. \code{~ treatment}).
\item \code{RCTtool.control}: String of control arm.
\item \code{RCTtool.xmod}: A list of one-sided formulas with covariates
on the right-hand side of the equation.
See \code{\link[=rct_lm]{rct_lm()}} in detail.
\item \code{RCTtool.plot_family}: The name of the font family
used to output the figure with {ggplot2}.
\item \code{RCTtool.table_fontsize}: The font size in the table
output by {kableExtra} and {flextable} used through {modelsummary}.
}
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
