% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.r
\name{summary.RCTtoolbox.power.analysis}
\alias{summary.RCTtoolbox.power.analysis}
\title{Summary of power analysis}
\usage{
\method{summary}{RCTtoolbox.power.analysis}(object, ...)
}
\description{
Summary of power analysis
}
