% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_ttest.r
\name{ttest}
\alias{ttest}
\title{Method of RCTtoolbox: \code{ttest()}}
\arguments{
\item{ctrl}{character vector. New control arm.
If NULL (default),
the first element of \code{treat_levels} is control arm.}

\item{subset}{subset condition.
If NULL (default),
full observations of \code{data} is used to implement t-test.}

\item{weights}{weight variable.
If NULL (default),
unbiased unweighted average and std.err. is calculated.}

\item{bootse}{integer.
Number of bootstrap samples to calculate std.err. of mean.
If 0 (default),
theoretical std.err. of mean is calculated.}

\item{bootp}{integer.
If 0 (default), welch two-sided t-test is implemented.
If 1 or more, the permutation test is implemented
(the number of permutations that randomly reallocate assignment
to keep the sample size collected from each group).}

\item{seed}{integer.
Seed value passed to the argument of \code{\link{set.seed}}.}
}
\value{
R6 object with "RCTtoolbox.ttest" class.
The returned object has following field and methods:
\describe{
\item{\code{data}}{Field.
Data frame including estimated result.
See the section "Data Field."}
\item{\code{print()}}{Method.
Print information about the returned object.
Run \code{$print()}.}
\item{\code{plot()}}{Method.
Visualization of result.
Run \code{$plot()}.}
\item{\code{summary()}}{Method.
Print result in console.
Run \code{$summary()}.}
}
}
\description{
When you create R6 object "RCTtoolbox"
via \code{\link{create_RCTtoolbox}}
and run its method \code{ttest()},
the method internally implements a function
that performs two-sided t-test or permutation test.
The method \code{ttest()} provides some arguments.
See the section "Arguments."
}
\section{Data Field}{


The R6 object with "RCTtoolbox.ttest" class has \code{data} field.
This field has a data frame with 13 variables:
\describe{
\item{mean1}{Outcome average of each experimental arm}
\item{se1}{Std.err. of outcome average of each experimental arm}
\item{n1}{Number of observations of each experimental arm}
\item{mean0}{Outcome average of control arm}
\item{se0}{Std.err. of outcome average of control arm}
\item{n0}{Number of observations of control arm}
\item{diff}{Difference between "mean1" - "mean0"}
\item{t}{t-statistics when \code{bootp = 0L} (Two-sided t-test)}
\item{df}{degree of freedom of t-distribution
when \code{bootp = 0L} (Two-sided t-test)}
\item{pval}{p-value of difference mean test.
Pr(>|t|) when \code{bootp = 0L} (Two-sided t-test).
Pr(>|diff|) when \code{bootp > 0L} (Permutation test).}
\item{method}{Character. Method of difference mean test.}
\item{arms}{Factor. Experimental arms.
The first level is control arm.}
\item{outcome}{Character. Outcome variable.}
}
}

\examples{
\dontrun{
  data(RubellaNudge)
  rct <- create_RCTtoolbox(
    atest + avacc ~ treat,
    data = RubellaNudge,
    treat_levels = LETTERS[1:7]
  )

  # two-sided t-test
  rct$ttest()
  # permutation test with control arm "C"
  rct$ttest(bootp = 50, ctrl = "C")
}

}
