% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.r
\name{ttest}
\alias{ttest}
\title{T-Test for RCTs with Multiple Experimental Arms}
\usage{
ttest(mod, data, base = NULL)
}
\arguments{
\item{mod}{formula.
Specify \code{y ~ d} where \code{y} is outcome and \code{d} is treatment.}

\item{data}{data object you want to use.}

\item{base}{character. Specify the name of control group.
Default is NULL.
If \code{base = NULL}, the first level of \code{d} is the control.}
}
\value{
tibble frame containing with five variables.
\code{treat} is a factor of treatments (baseline is the control).
\code{mean} and \code{se} is a mean and its standard error of each group.
\code{effect} is a difference of mean between each group and the control.
\code{p.value} is a p-value of t-test.
}
\description{
This function performs a t-test to test the null hypothesis
that there is no difference between the means of the two groups.
The user specifies the outcome variable and treatment variable
in the formula object.
When there are three or more treatment groups,
this function tests the difference of mean between
the control group specified by the user with the \code{base} argument
and each group (including the control group).
}
\examples{
# DGP
set.seed(120511)
n <- 1000
x1 <- rnorm(n); x2 <- rnorm(n)
d <- sample(c("A", "B", "C"), size = n, replace = TRUE)

ya <- 0.2 + 0.5 * x1 + 0.01 * x2
yb <- 1.2 + 0.3 * x2
yc <- -1 - 0.2 * x1 + 0.5 * x2
y <- ifelse(d == "A", ya, ifelse(d == "B", yb, yc))

dt <- data.frame(y, d, x1, x2)

# t-test
ttest(y ~ d, dt)

}
