% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtool_ttest.r
\name{ttest}
\alias{ttest}
\title{RCTtoolbox: T-test}
\usage{
ttest(
  baseline,
  data,
  ctrl,
  subset = NULL,
  weights = NULL,
  bootse = NULL,
  bootp = NULL,
  seed = 120511
)
}
\arguments{
\item{baseline}{(list of) baseline formulas. \code{outcome ~ treatment}.}

\item{data}{data.frame/tibble object you want to use.}

\item{subset}{subset condition.}

\item{weights}{weight variable.}

\item{bootse}{numeric. Number of bootstrap sample to calculate se.
If NULL, se is calculated by effective sample size.}

\item{bootp}{numeric. Number of re-randomization
to conduct permutation test.
If NULL, perform (welch) two-sided t-test.}

\item{seed}{numeric. seed value.}
}
\description{
This function performs a t-test to test the null hypothesis
that there is no mean difference between the two groups.
When there are three or more treatment groups,
this function tests the difference of mean between
the control group specified by the user with the \code{ctrl} argument
and each group (including the control group).
}
\examples{
\dontrun{
data(RubellaNudge)
rct <- new_RCTtool(
  atest + avacc ~ treat,
  data = subset(RubellaNudge, coupon == 1),
  ctrl = "A"
)
rct$ttest()$result
rct$ttest(ctrl = "C", bootp = 50)$result
}

}
