% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTtoolbox_ttest.r
\name{ttest_multi_arm}
\alias{ttest_multi_arm}
\title{T-test for Multiple Experimental Arms}
\usage{
ttest_multi_arm(
  baseline,
  data,
  levels,
  labels = NULL,
  ctrl = NULL,
  subset,
  weights,
  bootse = NULL,
  bootp = NULL,
  seed = 120511
)
}
\arguments{
\item{baseline}{formula. \code{outcome ~ treatment}.}

\item{data}{data.frame/tibble object you want to use.}

\item{subset}{subset condition.}

\item{weights}{weight variable.}

\item{bootse}{numeric. Number of bootstrap sample to calculate se.
If NULL, se is calculated by effective sample size.}

\item{bootp}{numeric. Number of re-randomization
to conduct permutation test.
If NULL, perform (welch) two-sided t-test.}

\item{seed}{numeric. seed value.}
}
\description{
T-test for Multiple Experimental Arms
}
